(lex) Don't emit REM tokens from lexer
While it's nice to know I suppose, the parser won't do anything with
them and will just drop them.  There's really no need to make the parser
do that work later, so we can just drop them in the lexer.

If we ever need to handle magic comments (which we might for unit
testing BrightScript files!), we should be able to revert this commit.