(lex) Don't make separate spaced reserved words
While `end if` is technically a full statement in BrightScript, it's a
composition of the `end` reserved word and the `if` reserved word.  It's
the *parser*'s job to figure out the context for the `end`, so making
the lexer try to read both words and create a single reserved word is a
violation of that separation of concerns.

If this winds up being wrong, at least the original implementation of
that logic exists in history now!