(parse) Add basic parser & an AST-printing visitor
Now that we've got a working lexer, we can start parsing those tokens
into an abstract syntax tree (or a "parse tree", depending on how you
prefer your terminology).  Thanks to @Slushy's grammar, building up a
parser for simple arithmetic operations was pretty easy, and should be
pretty easy to extend later.

Next up: a whole lotta tests for this parser!  And figuring out an
appropriate way to test a parser!