(parse) Start testing parser
I'm not really sure if I like this pattern of building up an AST at
runtime and comparing it to an executed one.  It's just really verbose
and a pain to maintain.  Maybe something like Jest's
[snapshots](http://facebook.github.io/jest/docs/en/snapshot-testing.html)
would work here?  It matches what the TypeScript compiler does with
benchmarks.  There's a few mocha equivalents around too, so maybe those
would be better.