(exec) Evaluate a bunch of binary operators and groupings
That's _almost_ all we need for a really silly initial version
(basically just a calculator).  Things we still need:

1. Protection against `-"foo"`
2. Protection against `not "bar"`
3. Exponentiation operators
4. A h*cking bunch of tests

(4) above is interesting, because it seems like we'll just need to throw
a bunch of possibilities at the executioner and see what breaks it.

Also I'm starting to hate the "executioner" name.