(grammar) Move assignments to lower precedence in grammar
Since assignments in BrightScript take the place of declarations in most
other languages, they need to be lower precedence (i.e. appear in the
top-down grammar before statements).

Since this grammar isn't being validated or used anywhere this is quite
likely wrong and quite likely a moot effort, but I'm sticking with it
for now gosh darnit!