fix(lex): Read multi-word form of "else if" and "end *"
BrightScript treats "elseif" and "else if" equivalently, and does the
same for "endif"/"end if", "endwhile"/"end while", "endsub"/"end sub",
"endfunction"/"end function".  It makes lexing the language a little
harder, because we need to speculatively read the next word whenever we
read an identifier called `end`.  Perhaps there's a more efficient way
to do this -- maybe a pass on the lexed tokens that finds
`[..., "end", "if", ...]` and transforms it into `[..., "endif", ...]`?
I'm sure there's some precedent for that, but that also gets tricky.

For now, this is fine.  Just try to avoid variables named `end` for
mildly better lexer performance, I guess. :cry: