feat(lex,parse,interp): Implement while loops with exits
BrightScript 'while' loops can be exited early with `exit while`, which
means we need some way to keep track of why a block stopped exiting.
Sometimes there's an error, sometimes it's an `exit while` statement,
sometimes it's from a `return` statement!  That last one isn't supported
yet (because we don't have functions), but the point still stands.

This is a big moment, because BrightScript execution is no longer
proportional to the length of a `.brs` file.