feat(types): Add rudimentary call-time type checking
In order to ensure arguments passed to a function are the correct type
(as they are in the reference BrightScript implementation on Roku), we
need to know what types a function _should_ expect.  This means:

* A set of accepted arguments known when the function is implemented,
  including:
    * Each argument's name
    * Each argument's expected type (or `dynamic` as the default)
    * Each argument's default value (if any)
* The name of the function (unless it's an anonymous function)

We'll eventually need to know the return type of each function, but I
can't remember where BrightScript uses that at runtime.  Perhaps it's to
ensure a `sub` doesn't return a value?