fix(interp): Disallow array comparisons
It's currently impossible to compare roArray values in the Reference
BrightScript Implementation (RBI).  IT results in a "type mismatch"
error, even if the left-hand side and right-hand side are of the same
type -- or even if they're the same value!  For example, this causes a
type mismatch:

```brightscript
a = [ "foo" ]
if a = a then print "array identity checks work"
```