feat(stdlib): Add trig functions Atan, Cos, Sin, Tan (#71)
This is a simple mapping of brightscript trig functions to javascript
trig functions.  Both brightscript and javascript use radians so the
mapping is trivial.  

For unit tests I chose a couple interesting values and added some
end-to-end tests.  I also ran the end-to-end tests on a Roku and got
similar output.

fixes #45