feat(lex,preprocessor): Add basic support for conditional compilation (#120)
While there's no support for the `bs_const` field of a `manifest` file
yet (because there's no support for `manifest` files at all), this
covers the bulk of the work required to support conditional compilation.
Namely:

1. Lexing the previously-unsupported `#if`, `#else if`, `#else`, `#end
   if`, `#const` and `#error` tokens
2. Parsing those tokens separately into a pseudo-AST
3. Evaluating that AST based on `#const` directives to selectively
   remove tokens
4. Inserting this preprocessor before the "standard" parser, so that
   conditional compilation remains transparent to the standard parsing
   process

~There's one big question for now though:  do `#const` declarations
exist only on a per-file basis, or are they cross-file?  If so, we'll
need to run the preprocessor across *all* files (where the processing
order would be significant as well) before passing the filtered tokens
to the "standard" parser.~

~I'll have to check that against the reference implementation, but for
now this is a meaty chunk of logic.~

As it turns out, `#const` values are in-scope only for *the file in
which they're declared*.  Attempting to reference a `#const` from
another file without defining it in the current file results in a
compilation error.