fix(arch): Use babel to downlevel async/await for node6
Node 6 is still a supported version, but it doesn't support the
async/await keywords that make unit testing with Promises less of a
chore.  While it's a bummer to have two different transpilers
(TypeScript for source and Babel for tests), the Babel transpiler is
significantly faster than the TypeScript one is -- even for JavaScript
files.