fix(parse): Allow `next` keyword as synonym for `end for`
While not specified anywhere in its public documentation, the Reference
BrightScript Implementation (RBI) allows `next` to be used to terminate
`for` and `for each` loops everywhere `end for` would normally be used.
Note that `next` can't be used like `continue` would in other languages
(i.e. to stop execution of the loop's block for the current loop
variable and move to the next value) - it's a strict synonym for `end
for` in BrightScript.

fixes #145