WIP provide locations for preprocessor tokens
This is going to be really painful for the mainline parser. I may have
to consider generating some of this code on a "just once" basis to keep
these tests isolated from the lexer's implementation, because I _really_
don't feel like spending forever setting locations for every token in
`test/`.  But I only have to do it once, so maybe it's worth it?