fix(lexer): Follow estree/esprima Position conventions
The [estree](https://github.com/estree/estree) and
[esprima](https://github.com/jquery/esprima) projects use zero-based
column numbers with ranges that include one character past the last
column that a token spans.  It's a small difference, but it means the
start and end columns are appropriate to pass *directly* into
`String.prototype.substring`.  For example, consider the following
input:

```
' 2 4 6 8    (column numbers)
foo = 1 + 2
```

In the estree/esprima pattern, the identifier `foo` would
have location:

```
{
  start: { line: 2, column: 0 },
  end:   { line: 2, column: 3 }
}
```