feat(lex,parse): Add support for type designators (#203)
Adds parser support for type designators (i.e. a trailing `$` `%` `#`
`!` after a variable identifier). This will still need to be implemented
in the runtime at a later time.

All function, variable, and parameter identifier tokens are permitted to
end with a type designator. Then, in the parser, we enforce that
function identifiers are NOT allowed to end with a type designator.

fixes #162