feat(extension): Add _brs_.runInScope (#326)
While the standard `Run` function allows arbitrary files to be executed,
they're executed in their own empty scope!  That leads to a whole lot of
repeated lexing, parsing, etc. of the `source/` tree when running unit
tests, so we can expose a custom version of `Run` that keeps the
previous values in-scope, deletes the previous `main` function, and
otherwise executes the requested files the same way.