fix(dirs): Defer process.cwd() calls until needed
Settting `process.cwd()` when the `brs` module is imported causes it to
be static for the lifetime of a consuming process, and that's typically
not what we want!  Fetching `process.cwd()` as part of the `brs`
execution path allows the root directory for `pkg:/` paths to vary with
the consuming process dynamically.