feat(stdlib): No-op Library statements instead of throwing
Library statements are becoming pretty common, and it'd be nice to at
least let code with `Library` statements be executed — even if the
libraries themselves aren't available.  I'm still looking for an
approach that'll allow these to be implemented without copyright
concerns, so for now just warning that `Library` statements aren't
supported should be sufficient.

relates to #217 but doesn't fix it