feat(stdlib): Implement ifSGNodeField.hasField() (#368)
The implementation is simpler than the existing _getField_ method, which
returns a value, while this simply returns a boolean if the field is
present.  

There is at least one bug in hasField in the official implementation
that I'm aware of, where calling hasField() with specific strings will
actually create the field, but this is a bug and I don't think we should
match that for compatibility sake.  

resolves #367