feat(stdlib): Allow mocking of "unknown" functions (#442)
This change allows us to mock functions that don't exist in source. This
seems somewhat counter-intuitive: why would we let a test implement a
function that doesn't exist in source, yet gets called from source code?

The reason is because `brs` doesn't have all the built-in global
functions implemented yet, so this unblocks the testing of source code
that calls one of those not-yet-implemented functions. The example I was
running into was `tr` -- I wanted to test the code that called `tr`, but
`tr` isn't implemented, and it also wasn't mockable, so I had no way to
test it.

Also, maybe this would be useful for TDD?