chore(deps): Upgrade to prettier 2.x (#460)
The defaults changed for prettier -- mostly `arrowParens: true` -- and
I've been meaning to enable that anyway.  Let's use that major-version
bump as an excuse to reformat everything.  It pains me a little to have
such a large diff but at least it's mostly `()` insertion.

Realistically, the main motivation for the upgrade is support for
TypeScript 3.8 language features like `import type …`.