feat(rsg): Require a host node for observeField
The reference implementation doesn't allow `observeField` (and friends)
to be called without a "host node": the SceneGraph node that holds
the callback function whos name is passed to `observeField`.  In
other words, this is not allowed from the main entry point of a channel:

```brightscript
sub main()
    p = createObject("roSGNode", "Poster") ' works!
    p.observeField("loadStatus", "onLoadStatusChanged")
    ' => BRIGHTSCRIPT: ERROR: roSGNode.ObserveField: no active host node: pkg:/source/main.brs(3)
end sub

sub onLoadStatusChanged(event as object)
    ' does something with new field
end sub
```

The concept of a Host Node appears to be the key to making
`unobserveFieldScoped` work differently than `unobserveField`.  While
this commit doesn't include support for either of the `unobserveField*`
functions, it does make it significantly easier to build them.

supports #505