feat(scenegraph,lex,parse): Add support for <ComponentLibrary/>
The <ComponentLibrary/> component loads a set of namespaced components
from a location declared in its `uri` field.  Some remote resources can
be loaded, but they require signing and packaging -- both of which are
out of scope for this project.  Load `pkg:/`-local component libraries,
which rely only on `.zip` compression for packaging.

fixes #589