feat(rsg): Support adding new BrsObjects at Runtime (#583)
This is a series of changes that allow other software using BRS as a
library to add new BrsComponents or Scenegraph components (PinPad,
BusySpinner, etc) at Runtime.

This is useful, for example, if another piece of software wanted to
implement video playback or Draw2d functionality, and then it could do
so by implementing those components locally and still leverage the BRS
project.

While the BRS project is very useful as a parser/lexer/interpreter, it
does not do any emulation or real-world display of Scenegraph nodes. If
there were another piece of software the wanted to leverage BRS, but do
that display piece, then being able to override/extend the built in
implementation would be a good thing.